+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.0-next.3
+
+### Patch Changes
+
+- a70869e775: Updated dependency `msw` to `^0.43.0`.
+- Updated dependencies
+  - @backstage/core-plugin-api@1.0.4-next.0
+  - @backstage/core-components@0.10.0-next.3
+
 ## 0.2.0-next.2
 
 ### Patch Changes