+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-bazaar
 
+## 0.1.22-next.3
+
+### Patch Changes
+
+- 4e9a90e307: Updated dependency `luxon` to `^3.0.0`.
+- Updated dependencies
+  - @backstage/core-plugin-api@1.0.4-next.0
+  - @backstage/core-components@0.10.0-next.3
+  - @backstage/catalog-client@1.0.4-next.2
+  - @backstage/cli@0.18.0-next.3
+  - @backstage/catalog-model@1.1.0-next.3
+  - @backstage/plugin-catalog-react@1.1.2-next.3
+  - @backstage/plugin-catalog@1.4.0-next.3
+
 ## 0.1.22-next.2
 
 ### Patch Changes