+++ plugins/catalog-graph/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-graph
 
+## 0.2.19-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.0.4-next.0
+  - @backstage/core-components@0.10.0-next.3
+  - @backstage/catalog-client@1.0.4-next.2
+  - @backstage/catalog-model@1.1.0-next.3
+  - @backstage/plugin-catalog-react@1.1.2-next.3
+
 ## 0.2.19-next.2
 
 ### Patch Changes