+++ plugins/fossa/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-fossa
 
+## 0.2.39-next.3
+
+### Patch Changes
+
+- 322d1ceeba: Allow configuration of base URL for Fossa links
+- a70869e775: Updated dependency `msw` to `^0.43.0`.
+- 4e9a90e307: Updated dependency `luxon` to `^3.0.0`.
+- Updated dependencies
+  - @backstage/core-plugin-api@1.0.4-next.0
+  - @backstage/core-components@0.10.0-next.3
+  - @backstage/catalog-model@1.1.0-next.3
+  - @backstage/plugin-catalog-react@1.1.2-next.3
+
 ## 0.2.39-next.2
 
 ### Patch Changes