+++ plugins/todo-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-todo-backend
 
+## 0.1.31-next.2
+
+### Patch Changes
+
+- a70869e775: Updated dependency `msw` to `^0.43.0`.
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.3
+  - @backstage/catalog-client@1.0.4-next.2
+  - @backstage/integration@1.2.2-next.3
+  - @backstage/catalog-model@1.1.0-next.3
+
 ## 0.1.31-next.1
 
 ### Patch Changes